Move TargetIcon to dedicated location in @components
