+++ assets/js/components/TargetIcon/TargetIcon.stories.jsx
