+++ assets/js/components/TargetIcon/index.js
@@ -1,0 +1,3 @@
+import TargetIcon from './TargetIcon';
+
+export default TargetIcon;
